pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "blockcpi"
version = "0.1"
description = "Blokus for CPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
   "requests",
   "pillow",
]

[project.optional-dependencies]
dev = ["black", "flake8", 'isort', 'Flake8-pyproject']

[tool.flake8]
max-line-length = 120